firmware: Introduce reboot_reason driver
This patch adds a basic driver to allow for commands like
"reboot bootloader" and "reboot recovery" to communicate this
reboot-reason to the bootloader.

This is commonly done on Android devices, in order to reboot
the device into fastboot or recovery mode. It also supports
custom OEM specific commands, via "reboot oem-<value>".

This driver pulls the phys memory address from DT as well as
the magic reason values that are written to the address for
each mode.

Change-Id: Idf315076b3c5f6345bb9c2d5333627b9b22fe5c8
Cc: Vishal Bhoj <vishal.bhoj@linaro.org>
Cc: haojian.zhuang@linaro.org
Cc: guodong.xu@linaro.org
Signed-off-by: John Stultz <john.stultz@linaro.org>