dts: hisi3660-hikey960: fix reserved memory region
After Hisilicon reviewed the reserved memory regions, found the MCU
reserved region is incorrect. The old reserved regions are used by MCU
with FAMA enabled, but now we are using MCU with FAMA disabled. So
change reserved memory region to correct value corresponding to the MCU
firmware we are using.

Signed-off-by: Leo Yan <leo.yan@linaro.org>