arm64: rename poplar_android_defconfig to poplar_defconfig
People are used to build a kernel for Poplar board from poplar_defconfig
for both Linux and Android kernel.  Let's rename poplar_android_defconfig
to poplar_defconfig on poplar-android-4.9 branch, so that people can
consistently use poplar_defconfig to build kernel for both poplar-4.9
and poplar-android-4.9 branch.

As the result, the original poplar_defconfig on poplar-android-4.9
branch gets removed.  It's a good thing anyway, as that defconfig
only works for poplar-4.9 branch.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>